BJS2-65863-write-recommendation FIX: create RecommendationCreateDto, RecommendationViewDto, SkillOfferViewDto,SkillOfferCreateDto. fix problems with anyDto,validation method in Service,

+++ src/main/java/school/faang/user_service/service/RecommendationService.java
@@ -8,8 +8,10 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
-import school.faang.user_service.dto.recommendation.RecommendationDto;
-import school.faang.user_service.dto.skilloffer.SkillOfferDto;
+import school.faang.user_service.dto.recommendation.RecommendationCreateDto;
+import school.faang.user_service.dto.recommendation.RecommendationViewDto;
+import school.faang.user_service.dto.skilloffer.SkillOfferCreateDto;
+import school.faang.user_service.dto.skilloffer.SkillOfferViewDto;
 import school.faang.user_service.entity.Skill;
 import school.faang.user_service.entity.User;
 import school.faang.user_service.entity.UserSkillGuarantee;
@@ -24,9 +26,38 @@ import school.faang.user_service.repository.recommendation.RecommendationReposit
 import school.faang.user_service.repository.recommendation.SkillOfferRepository;
 
 import java.time.LocalDateTime;
-import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
+import java.util.stream.StreamSupport;
 
+/**
+ * Сервис для управления рекомендациями пользователей.
+ * <p>
+ * Этот сервис предоставляет методы для создания, обновления, удаления и получения рекомендаций,
+ * а также для управления навыками, предлагаемыми в рекомендациях.
+ * </p>
+ * <p>
+ * Основные функции:
+ * <ul>
+ *     <li>{@link #create(RecommendationCreateDto) Создание новой рекомендации} с проверкой валидности данных.</li>
+ *     <li>{@link #update(RecommendationCreateDto) Обновление существующей рекомендации}.</li>
+ *     <li>{@link #delete(long) Удаление рекомендации} по её идентификатору.</li>
+ *     <li>{@link #getAllUserRecommendations(long) Получение списка всех рекомендаций}, полученных пользователем.</li>
+ *     <li>{@link #getAllGivenRecommendations(long) Получение списка всех рекомендаций}, созданных пользователем.</li>
+ *     <li>{@link #validateRecommendation(RecommendationCreateDto) Проверка валидности рекомендации}, включая проверку временных интервалов и существования навыков.</li>
+ *     <li>{@link #saveSkillsOffer(RecommendationCreateDto) Сохранение предложенных навыков} в рекомендации и добавление гарантов к навыкам пользователя.</li>
+ * </ul>
+ * </p>
+ * @author marsel_mkh
+ * @see RecommendationViewDto
+ * @see RecommendationCreateDto
+ * @see SkillOfferViewDto
+ * @see Recommendation
+ * @see SkillOffer
+ * @see User
+ * @see Skill
+ * @see UserSkillGuarantee
+ */
 @Slf4j
 @Data
 @Service
@@ -46,17 +77,16 @@ public class RecommendationService {
      * @param recommendation  объект DTO с данными о рекомендации
      * @return RecommendationDto - созданная рекомендация
      */
-    public RecommendationDto create(@NonNull RecommendationDto recommendation) {
+    public RecommendationViewDto create(@NonNull RecommendationCreateDto recommendation) {
         validateRecommendation(recommendation);
 
         long receiverId = recommendation.getReceiverId();
-        User user = getUser(receiverId);
+        User receiver = getUser(receiverId);
 
         long authorId = recommendation.getAuthorId();
         User author = getUser(authorId);
-
-        Recommendation recommendationEntity = recommendationMapper.toEntity(recommendation);
-        recommendationEntity.setReceiver(user);
+        Recommendation recommendationEntity = recommendationMapper.CreateDtoToEntity(recommendation);
+        recommendationEntity.setReceiver(receiver);
         recommendationEntity.setAuthor(author);
         List<SkillOffer> skillOffers = SkillOffersToEntity(recommendation);
         recommendationEntity.setSkillOffers(skillOffers);
@@ -71,7 +101,7 @@ public class RecommendationService {
      * @param recommendation - объект DTO с обновленными данными
      * @return RecommendationDto - обновленная рекомендация
      */
-    public RecommendationDto update(@NonNull RecommendationDto recommendation) {
+    public RecommendationViewDto update(@NonNull RecommendationCreateDto recommendation) {
         validateRecommendation(recommendation);
         long receiverId = recommendation.getReceiverId();
         long authorId = recommendation.getAuthorId();
@@ -81,7 +111,6 @@ public class RecommendationService {
         skillOfferRepository.deleteAllByRecommendationId(recommendation.getId());
 
         saveSkillsOffer(recommendation);
-
         Recommendation recommendationEntity = recommendationRepository.findById(recommendation.getId()).
                 orElseThrow(() -> {
                     log.error("Рекомендация с ID {} не найдена", recommendation.getId());
@@ -94,14 +123,14 @@ public class RecommendationService {
     /**
      * Удаляет рекомендацию по ее идентификатору
      * @param recommendationId - идентификатор рекомендации
+     * @throws DataValidationException если рекомендация не найдена
      */
     public void delete(long recommendationId) {
-        recommendationRepository.findById(recommendationId).orElseThrow(() -> {
-            log.error("Рекомендация с ID {} не найдена", recommendationId);
-            return new DataValidationException(String.format("Recommendation id %d not found", recommendationId));
-        });
-
-        recommendationRepository.deleteById(recommendationId);
+        if (recommendationRepository.existsById(recommendationId)) {
+            recommendationRepository.deleteById(recommendationId);
+        }
+        log.error("Рекомендация с ID {} не найдена", recommendationId);
+        throw new DataValidationException(String.format("Recommendation id %d not found", recommendationId));
     }
 
     /**
@@ -109,7 +138,7 @@ public class RecommendationService {
      * @param receiverId - идентификатор пользователя
      * @return List<RecommendationDto> - список полученных рекомендаций
      */
-    public List<RecommendationDto> getAllUserRecommendations(long receiverId) {
+    public List<RecommendationViewDto> getAllUserRecommendations(long receiverId) {
         Pageable pageable = PageRequest.of(0, Integer.MAX_VALUE);
         List<Recommendation> recommendationPage =
                 recommendationRepository.findAllByReceiverId(receiverId, pageable).toList();
@@ -123,7 +152,7 @@ public class RecommendationService {
      * @param authorId - идентификатор автора
      * @return List<RecommendationDto> - список данных о рекомендациях
      */
-    public List<RecommendationDto> getAllGivenRecommendations(long authorId) {
+    public List<RecommendationViewDto> getAllGivenRecommendations(long authorId) {
         Pageable pageable = PageRequest.of(0, Integer.MAX_VALUE);
         List<Recommendation> recommendationPage = recommendationRepository
                 .findAllByAuthorId(authorId, pageable).toList();
@@ -137,12 +166,14 @@ public class RecommendationService {
      * чем через 6 месяцев после его последней рекомендации этому пользователю.
      * Также проверяет, что предлагаемые в рекомендации навыки существуют в системе.
      * @param recommendation - объект DTO с рекомендацией
+     * @throws DataValidationException если рекомендация не прошла валидацию
      */
-    void validateRecommendation(RecommendationDto recommendation) {
+    private void validateRecommendation(RecommendationCreateDto recommendation) {
         LocalDateTime sixMonthsAgo = LocalDateTime.now().minusMonths(6);
-        LocalDateTime recommendationDate = recommendation.getUpdatedAt();
+        LocalDateTime recommendationUpdateAt = getUpdateAtCreateRecommendation(recommendation)
+                .orElseThrow(() -> new DataValidationException("Update date is not found"));
 
-        if (recommendationDate.isBefore(sixMonthsAgo)) {
+        if (recommendationUpdateAt.isBefore(sixMonthsAgo)) {
             log.error("Ошибка валидации: рекомендация обновлена слишком рано");
             throw new DataValidationException("Updated recommendation too early");
         }
@@ -152,38 +183,50 @@ public class RecommendationService {
             throw new DataValidationException("Skill offer is not found");
         }
 
-        List<SkillOfferDto> skillOffers = recommendation.getSkillOffers();
+        List<SkillOfferCreateDto> skillOffers = recommendation.getSkillOffers();
         List<Long> skillIds = skillOffers.stream()
-                .map(SkillOfferDto::getSkillId)
+                .map(SkillOfferCreateDto::getSkillId)
                 .toList();
         for (Long skillId : skillIds) {
-            skillOfferRepository.findById(skillId).orElseThrow(() -> {
+            if (!skillRepository.existsById(skillId)) {
                 log.error("Навык с ID {} не найден в системе", skillId);
-                return new DataValidationException("Skill offer not found");
-            });
+                throw new DataValidationException("Skill offer not found");
+            }
         }
+    }
+    private Optional<LocalDateTime> getUpdateAtCreateRecommendation(RecommendationCreateDto recommendation) {
+        User recommendationAuthor  = getUser(recommendation.getAuthorId());
+        return recommendationAuthor.getRecommendationsGiven().stream()
+                .filter(rec -> {
+                    User receiver = rec.getReceiver();
+                    User createRecommendationReceiver = getUser(receiver.getId());
+                    return receiver.equals(createRecommendationReceiver);
+                })
+                .findFirst()
+                .map(Recommendation::getUpdatedAt);
     }
 
     /**
-     * Сохранить предложенные в рекомендации скиллы в репозиторий SKillOfferRepository используя его метод create.
+     * Сохраняет предложенные в рекомендации навыки в репозиторий {@link SkillOfferRepository} используя его метод {@code create}.
      * Если у пользователя, которому дают рекомендацию, такой скилл уже есть,
-     * то добавить автора рекомендации гарантом к скиллу, который он предлагает,
+     * то добавляет автора рекомендации гарантом к скиллу, который он предлагает,
      * если этот автор еще не стоит там гарантом.
-     * @param recommendation  рекомендации
+     * @param recommendation рекомендация
+     * @throws DataValidationException если навык не найден
      */
-    private void saveSkillsOffer(RecommendationDto recommendation) {
+    private void saveSkillsOffer(RecommendationCreateDto recommendation) {
         long receiverId = recommendation.getReceiverId();
         long authorId = recommendation.getAuthorId();
 
         User receiver = getUser(receiverId);
         User author = getUser(authorId);
 
-        for (SkillOfferDto skillOfferDto : recommendation.getSkillOffers()) {
+        for (SkillOfferCreateDto skillOfferDto : recommendation.getSkillOffers()) {
             long skillId = skillOfferDto.getSkillId();
             skillOfferRepository.create(skillId, recommendation.getId());
 
             Skill skill = skillRepository.findById(skillId).orElseThrow(() -> {
-                log.error("Ошибка: навык с ID {} не найден", skillId);
+                log.error("Ошибка: навык с ID {} в skillRepository не найден", skillId);
                 return new DataValidationException("Skill not found");
             });
 
@@ -213,6 +256,7 @@ public class RecommendationService {
      * Получает пользователя по его идентификатору
      * @param userId - идентификатор пользователя
      * @return User - найденный пользователь
+     * @throws DataValidationException если пользователь не найден
      */
     private User getUser(long userId) {
         return userRepository.findById(userId).orElseThrow(() -> {
@@ -225,16 +269,13 @@ public class RecommendationService {
      * Получает список предложенных навыков в рекомендации
      * @param recommendation - объект DTO с рекомендацией
      * @return List<SkillOffer> - список предложенных навыков
+     * @throws DataValidationException если навык не найден
      */
-    private List<SkillOffer> SkillOffersToEntity(RecommendationDto recommendation) {
-        List<SkillOffer> allSkillOffer = new ArrayList<>();
-        for (SkillOfferDto skillOfferDto : recommendation.getSkillOffers()) {
-            SkillOffer skillOffer = skillOfferRepository.findById(skillOfferDto.getId()).orElseThrow(() -> {
-                log.error("Ошибка: навык с ID {} не найден", skillOfferDto.getId());
-                return new DataValidationException("Skill not found");
-            });
-            allSkillOffer.add(skillOffer);
-        }
-        return allSkillOffer;
+    private List<SkillOffer> SkillOffersToEntity(RecommendationCreateDto recommendation) {
+        List<Long> skillOfferIds = recommendation.getSkillOffers().stream()
+                .map(SkillOfferCreateDto::getSkillId)
+                .toList();
+        Iterable<SkillOffer> allSkillOffers = skillOfferRepository.findAllById(skillOfferIds);
+        return StreamSupport.stream(allSkillOffers.spliterator(), false).toList();
     }
 }
